+++ lib/restart-trigger.ts
@@ -1,2 +1,2 @@
-// restart triggered at 2026-03-08T10:24:58.217Z
-export const RESTART_TS = 1772965498217;
+// restart triggered at 2026-03-08T17:54:44.049Z
+export const RESTART_TS = 1772992484049;